Add fluent API and refactor
Close #9
Close #10
